# packages/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: FuelEU Control Tower API
  version: 1.0.0
  description: OpenAPI Specification for FuelEU Maritime Compliance Platform ensuring deterministic generation for the UI Next.js client.
servers:
  - url: http://localhost:8082
    description: Local development server
paths:
...
  /api/v1/compliance-ledger/rows:
    get:
      summary: List compliance ledger rows for vessels
      operationId: listComplianceLedgerRows
      parameters:
        - in: query
          name: year
          required: false
          schema:
            type: integer
            format: int32
            default: 2025
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceLedgerRow'
  /api/v1/compliance-ledger/rows/{vesselId}:
    put:
      summary: Update a vessel ledger row for testing/simulation overrides
      operationId: updateComplianceLedgerRow
      parameters:
        - in: path
          name: vesselId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: year
          required: false
          schema:
            type: integer
            format: int32
            default: 2025
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceLedgerRowUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceLedgerRow'
        '400':
          description: Bad Request
  /api/v1/compliance-ledger/summary:
    get:
      summary: Retrieve compliance ledger summary KPIs
      operationId: getComplianceLedgerSummary
      parameters:
        - in: query
          name: year
          required: false
          schema:
            type: integer
            format: int32
            default: 2025
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceLedgerSummary'
...
components:
  schemas:
...
    ComplianceLedgerRow:
      type: object
      properties:
        vesselId:
          type: string
          format: uuid
        imo:
          type: string
        name:
          type: string
        vesselType:
          type: string
        buildYear:
          type: integer
          format: int32
          nullable: true
        flagState:
          type: string
        energyInScope:
          type: number
          format: double
          description: Energy in scope in MJ.
        actualIntensity:
          type: number
          format: double
          description: Actual GHG intensity in gCO2eq/MJ.
        targetIntensity:
          type: number
          format: double
          description: Regulatory target GHG intensity in gCO2eq/MJ.
        icb:
          type: number
          format: double
          description: Initial Compliance Balance in gCO2eq.
        vcb:
          type: number
          format: double
          nullable: true
          description: Verified Compliance Balance in gCO2eq.
        borrowedAmount:
          type: number
          format: double
          description: Borrowed compliance balance in gCO2eq.
        bankedAmount:
          type: number
          format: double
          description: Banked compliance balance in gCO2eq.
        reportingYear:
          type: integer
          format: int32
        status:
          type: string
          description: Deficit or Compliant.
    ComplianceLedgerSummary:
      type: object
      properties:
        totalIcb:
          type: number
          format: double
          description: Fleet aggregate ICB in gCO2eq.
        totalAcb:
          type: number
          format: double
          description: Fleet aggregate ACB/VCB proxy in gCO2eq.
        totalBorrowingCap:
          type: number
          format: double
          description: Fleet aggregate borrowing cap in gCO2eq.
        penaltyExposure:
          type: number
          format: double
          description: |
            Estimated FuelEU penalty exposure in EUR.
            Computed using Annex IV Part B:
            |complianceBalance[gCO2eq]| / (actualIntensity[gCO2eq/MJ] * 41,000[MJ/tVLSFOeq]) * 2,400[EUR/tVLSFOeq].
        rowCount:
          type: integer
          format: int32
        year:
          type: integer
          format: int32
    ComplianceLedgerRowUpdateRequest:
      type: object
      properties:
        energyInScope:
          type: number
          format: double
          description: Override energy in scope (MJ).
        actualIntensity:
          type: number
          format: double
          description: Override actual intensity (gCO2eq/MJ).
        targetIntensity:
          type: number
          format: double
          description: Override target intensity (gCO2eq/MJ).
        icb:
          type: number
          format: double
          description: Override initial compliance balance (gCO2eq).
        vcb:
          type: number
          format: double
          nullable: true
          description: Override verified compliance balance (gCO2eq).
        bankedAmount:
          type: number
          format: double
          description: Override banked amount for test workflows (gCO2eq).
        borrowedAmount:
          type: number
          format: double
          description: Override borrowed amount for test workflows (gCO2eq).
...